requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
SQLAlchemy==2.0.38
pydantic==2.10.6
pymysql==1.1.1
cryptography==43.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
